shreyas/backend/requirements.txt
fastapi
uvicorn
pydantic
python-jose
passlib[bcrypt]
openai
pinecone
google-genai
python-dotenv
google-generativeai
pdfplumber
PyMuPDF
pypdf
pillow
tenacity
langchain-community
langchain-text-splitters
langchain-core
rank-bm25
numpy
chromadb

